Updated tileProperties
Now the names of the properties should be easier to understand, and the values are more consistent.
Changed the name of the type for conveyorbelts to conveyorType, and added a new type, which is always "conveyorBelt".
Added properties to starting points and pushers.
Changed the old directions in conveyorbelt from "up" to NORTH to be more consistent + reflect the Direction enum class